Update sidebar components
