+++ tv-shows/src/app/layout.tsx
@@ -1,8 +1,7 @@
 import type { Metadata } from "next";
 import { Providers } from "./providers";
-import { Box, Grid, Hide } from "@chakra-ui/react";
-import { SidebarNavigation } from "@/components/shared/Sidebar/SidebarNavigation";
-import { MobileHeader } from "@/components/shared/MobileHeader/MobileHeader";
+import { Grid, Hide } from "@chakra-ui/react";
+import { Sidebar } from "@/components/shared/Sidebar/Sidebar";
 import "./globals.css";
 
 export const metadata: Metadata = {
@@ -19,7 +18,6 @@ export default function RootLayout({
     <html lang="en">
       <body>
         <Providers>
-          <MobileHeader />
           <Grid
             templateColumns={{
               base: "1fr",
@@ -27,9 +25,7 @@ export default function RootLayout({
             }}
             bg="purple.500"
           >
-            <Hide below="md">
-              <SidebarNavigation />
-            </Hide>
+            <Sidebar />
             {children}
           </Grid>
         </Providers>
